задание 5.1
