+++ js/number-generate.js
@@ -1,0 +1,18 @@
+const getRandomN = (a, b) => {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandom = (min, max) => {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  return Math.floor(Math.random() * (max - min + 1) + min);
+}
+
+const getRandomElement = (elements) => elements[getRandomN(0, elements.length - 1)];
+
+const getRandomInteger = () => Math.floor(Math.random() * 100);
+
+export {getRandomN, getRandom, getRandomElement, getRandomInteger};
